Add Question 5 to logic flow
Add question to check if the user wants to add any deductions.
This uses the `multiple_choice` question type (radio buttons).

This is an optional question as the user may not have any
deductions to add. The hint text contains more information on
what deductions are applicable and should help the user decide
whether to answer this question.

Co-authored-by: Deborah Chua <deborah.chua@digital.cabinet-office.gov.uk>