Implement ChildBenefitTaxCalculator class
Add the calculator class and store variables from the response
from each question in the flow.

In the same way as other smart answers we have stored the tax years
with their start and end dates in a .yml file which can be accessed
through the `self.child_benefit_data` method.

In `Q3b` and `Q3d`, to store part year claim dates for multiple
children we can add them to a hash hash where we pass in the type
(a `start_date` or `stop_date` along with the `response`.

Co-authored-by: Deborah Chua <deborah.chua@digital.cabinet-office.gov.uk>