Add results page
Add page to display the amount of tax to pay and benefit amount
received.

We have also added some methods which are used to present some of
the data nicely for the view.

Only one results view is required as some of the text displayed
is dependent on the `tax_year` selected and the `calculate_adjusted_net_income`

Co-authored-by: Deborah Chua <deborah.chua@digital.cabinet-office.gov.uk>