Ask EU citizens 'When are you coming to the UK?' for clear results
Adds a new question for EU citizens: 'When are you coming to the
UK?'. If they answer 'before 1 January 2021', take them to an
outcome that says they do not need a visa.

If citizens answer 'from 1 January 2021', we will route them to
the existing flow for non-visa nationals, for example Argentina:
https://www.gov.uk/check-uk-visa/y/argentina

NB this commit also removes a failing Croatia test which adds no
value.

Trello: https://trello.com/c/WoFhTMCi/2196-5-1-dec-logic-changes-for-eu-citizens-check-if-you-need-a-uk-visa